requirements.txt
python-telegram-bot==21.0
Pillow==10.2.0
